voice-agent/requirements.txt
livekit-agents>=1.2.0
livekit-plugins-openai>=1.2.0
livekit-plugins-silero>=1.2.0
livekit-plugins-cartesia>=1.2.0
openai>=1.12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
